+++ documentation/Changelog.md
@@ -8,10 +8,13 @@ As of version 0.8.7, *Time Copy* was transformed into a sidebar-based interface.
 ---
 
 ## Last important changes summary
--
+- Fix filter not accepted after change
+
 ## V 1.3.22
 * **Fixes**
 - Show platform module list on start screen again
+- Catch undefiened errors from filters and platforms
+- Fix filter not accepted after change
 
 * **Other Changes* **
 - code cleanup (also storage changes)